Adding support for setting common virtualhost directives
This module is now able to set its own configuration options within the default virtualhost file without needing to re-template out the file.

This means this role will only need to worry about its own configuration options (minimal) and not the entire virtualhost file. Its options are 'injected' at the end of the main file as needed.

This keeps configuration files as simple as possible and removes the need for this role to track the 'upstream' apache module as it changes.